// frontend/src/components/model/CategoryEnum.ts
export type CategoryEnum =
    | "ANIMALS"
    | "ART"
    | "BRANDS"
    | "FICTIONAL_CHARACTERS"
    | "FOOD_AND_DRINK"
    | "GENERAL"
    | "GEOGRAPHY"
    | "HISTORY"
    | "LITERATURE"
    | "MATHEMATICS"
    | "MOVIES_AND_TV"
    | "MUSIC"
    | "NATURE"
    | "OBJECTS"
    | "OCCUPATIONS"
    | "POLITICS"
    | "SCIENCE"
    | "SPORTS";

export type NullableCategoryEnum = CategoryEnum | "";

export const ALL_CATEGORIES: CategoryEnum[] = [
    "ANIMALS",
    "ART",
    "BRANDS",
    "FICTIONAL_CHARACTERS",
    "FOOD_AND_DRINK",
    "GENERAL",
    "GEOGRAPHY",
    "HISTORY",
    "LITERATURE",
    "MATHEMATICS",
    "MOVIES_AND_TV",
    "MUSIC",
    "NATURE",
    "OBJECTS",
    "OCCUPATIONS",
    "POLITICS",
    "SCIENCE",
    "SPORTS"
];